✨ feat(schema): Add `Meta` class and `function_call` field
Added a new class `Meta` in the `schema.py` file to handle metadata related to the API. The `Meta` class has a boolean field `docs` to indicate whether the API is for documentation purposes.

Also, included a new field `function_call` in the `Message` class to store AI generated function calls. This field is optional and can be used to provide additional information about the API.

This commit also includes some refactoring and code cleanup.