:sparkles: feat(sender): add File.upload_file method
This commit adds a new method `File.upload_file` to the `sender` module. This method is used to upload a file in the `telegram`, `slack`, `kook`, and `discord` senders. The method takes two parameters: `name` and `data`, representing the name of the file and the file data respectively. It returns a `File` object.

The method is used in the `telegram`, `slack`, `kook`, and `discord` senders to upload files. The existing code that used `RawMessage.upload_file` has been updated to use the new method.

The changes in this commit include:

- Added `File` import in `llmkira/sender/telegram/__init__.py`, `llmkira/sender/slack/__init__.py`, `llmkira/sender/kook/__init__.py`, and `llmkira/sender/discord/__